backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
youtube-transcript-api==0.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
spacy==3.7.2
nltk==3.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
reportlab==4.0.7
markdown==3.5.1
